Clean up some droppings, update README
There are probably more droppings that Greg didn't remove from his tree
but could have.  I just happen to know these aren't used anymore.  :)

Changed the README to push people toward the use of fakeroot to build
the Debian package.  Using sudo makes it so that git cannot modify your
working directory.  To fix this before you pull, run:

   sudo chown -r <you> pkg

Obviously replacing <you> with your username.  Actually, if your working
directory is clean (git status), you can:

   sudo rm -rf pkg
   git checkout -f

That'll rebuild the pkg directory properly.  To avoid the problem in the
future, use fakeroot rather than sudo to make package and the system
will just pretend to chown everything as root for the purposes of
running dpkg-deb on it.  In future, we'll set up a proper debian/ dir
and the make pkg line will use dpkg-buildpackage or gbp as necessary to
build the package.

This should bring us up to something that can be merged.