Fix some uninitialized variables
In Applewin.cpp, that's genunely an un-initialized variable.  Arlo
probably commented out the ptm assignment not knowing immediately what
to do with it and planning to come back later.

For ftpparse.cpp, t should never be returned outside the for loop which
assigns it, but it hurts nothing to initialize t to silence the warning.