Add a shim for eventual porting to SDL2
Added a shim.c and shim.h where we can stuff code that isn't ours, isn't
supposed to be around forever, and which helps us port this mess to
SDL2.  Eventually this is where code from SDLCL will go.  In the
meantime, it contains the UNIX version of SDL_filesystem.c and the
appropriate header content to match, modified only insofar as they need
to be in order to compile in our source tree.

Didn't bother to match style with the rest of the code for this pretty
much on purpose as if SDL updates theirs before we stop using ours, I
intend to pick up their changes verbatim.

These functions aren't used yet—I'd have to look up where they need to
be used and I have to leave for a bit, so if Chris or David wants to
have a go at it now, they're welcome to.  Otherwise I'll do something
with it later tonight or tomorrow.