Load assets (resources) from BasePath
First, renamed RESOURCE_*_BMP to ASSET_*_BMP because resource tends to
mean specific things on gooey OSes.

Secondly, we now use SDL_GetBasePath() and store it in asset_basepath.
This means we always know where the linapple binary is located,
irrespective of where it's run from.  And that means we can load all of
our assets relative to that directory!  (Fixes #10)

Currently we just die if SDL_GetBasePath() fails.  Theoretically we
could fall back on checking . for those files.  Should we?