NULL IS NOT 0!
Fixed a few "not_pointer != NULL || false" constructs.  First, you never
need to "|| false" anything to indicate that something false if it is
not true, because if it is not true, by definition, it's false.

In several of these cases it looks like we're honestly comparing macro
constants we've #define'd with 0 and returning true regardless, but
whatever for the moment since maybe that's part of the brutal port to
Linux of a Windows emulator and maybe if we support emulating the right
hardware in the future it might not always be hardcoded.

I put the assignment of a comparison in parentheses for readability.
Didn't bother for the returns.