Redo how assets are loaded/managed
This is a first pass at centralizing the asset management.  Probably all
or nearly all of these surfaces are intended to be loaded from disk,
converted to screen pixel format, the disk format surface freed, and the
converted format saved.  That'd make a good second pass.  It doesn't
even bother me that this sounds like something that should be in the
video handling code because SDL 2.0 doesn't care.  The thing is, this
needs to be done after SDL_SetVideoMode, and right now it isn't.  So
pass #2 will have to wait a bit.