Extra safety to assets, fix dumb typo
Fixed dumb typo that inverted logic—figures right before bed.

While I was at it, we now check the return from Asset_Init and exit with
an error if it's false—it prints its own error messages.  And while I
was att it, I made assets itself a pointer that starts out null.  This
might be all totally unnecessary once we start cleaning up other things,
but this is in an early enough state I'd like us to have options.