Make make install/uninstall work
Cleaned up make install which didn't work, and recreated make uninstall
which didn't exist.  This undoes some of Greg's dir/* magic processing
in terms of assets at least.  That's probably okay though because that
kind of magic processing usually winds up breaking in some way, be it
major or minor.  For example we shouldn't be installing font.xcf.

Things in res/ which _should_ be installed are listed in INSTASSETS now,
no subdir needed.  That'll do for now.  The things presently in this
list will eventually be installed somewhere in $PREFIX/share, which is
why linapple.conf isn't here.

CONFFILES are also currently found in res/ and are just linapple.conf
for now.

These files all still get copied to bin on make, but the versions in bin
are not the ones that get installed.  That whole TARGETDIR thing is part
of Greg's package building.  I'm sort of planning to move this tree to
autoconf possibly and automake, on the assumption that a headache now
porting it will give us a lot of portability later to BSD, Mac, and even
Windows if built under MSYS2.  (Although builds under Windows probably
are going to want to use Visual Studio, and someone's gonna want XCode
on the mac…  These aren't my problem.)

Haven't touched make package yet.