Remove unnecessary extra files
We had two copies of font.xcf.  We only need one and arguably res is the
wrong place to put it.  The root is ideally not hte right place to put
it either, but we'll keep that one for the time being.

linapple.installed.conf isn't installed, and we have a linapple.conf
that is.  Get rid of it.

Printer.txt is empty.  The file should be created in the user's home
directory when they print something.