Load bmp files and Master.dsk from "proper" dir
It's customary on Windows to load asset files from the same directory as
the executable, but this is not normal for an installed Linux program.
We now try to load the asset files from ${PREFIX}/share/linapple/ first
and fall back on SDL_GetBasePath() if that doesn't work.  The former can
be overridden by changing ASSET_DIR when compiling.

For now both ASSET_DIR at compile time and SDL_GetBasePath at runtime
have fallbacks on ./ if they're not provided.  When we autoconf this
mess, the fallback for failure to define ASSET_DIR may go away.
Currently none of these allow SDL_GetPrefPath() to override them.  They
all could.  Do we want that?

At this point the only file still handled by existing code is the
registry file.  I need some help with deciding how we should handle
that.  Mostly it's complicated by the fact that the registry file
includes its own documentation.  If it were a simple ini file with
documentation in a text document or manpage or something, it'd be real
easy: All values would have defaults, the file we put in /etc would list
all of those default values and allow the sysadmin to override the
defaults, and we'd allow a user registry file to override those.  Not so
easy with the docs in the monolithic file not set up for override.