LoadAllConfigurations(): When making config dir/file, use XDG config dir
If no writable user configuration file is found, we create one. However,
this was incorrectly hardcoded to create linapple/linapple.conf under
$HOME/.config/; that directory is correct only if $XDG_CONFIG_HOME is not
set or is set to that value.

We have already calculated the correct location of the XDG user config dir,
so we can just re-use that value to fix this.

Also change the warning when we cannot find any of the environment
variables we needed to locate/create a user config file to make it clear
that we checked $XDG_CONFIG_HOME as well as $HOME.