Makefile: Remove unused RESOURCE_INIT_DIR
It's not clear what a "resource" was meant to be, given that we
already have ASSET_DIR for graphical and disk images and the like.

The comment mentions that `linapple.conf` would be in the resource
dir, but that's clearly a configuration file. Further, no default
version of that is necessary since the as-built default configuration
is contained in the binary itself. We should probably be supplying
only a `linapple.conf.sample` as documentation; this would help avoid
issues with it, `res/linapple.conf` (which appears to be unused) and
the internal default configuration getting out of sync. (Sysadmins
overriding the default configuration in `/etc/linapple/linapple.conf`
should be setting only their changes so their file doesn't
accidentally override changes to the default config in new versions of
LinApple.)

Anyway, it's best we remove this right now and re-implement it with
clearer naming and code if we decide in the future we need something
other than ASSET_DIR.